+++ mssqlcli/mssqlclioptionsparser.py
@@ -11,8 +11,6 @@ MSSQL_CLI_SERVER = u'MSSQL_CLI_SERVER'
 MSSQL_CLI_ROW_LIMIT = u'MSSQL_CLI_ROW_LIMIT'
 MSSQL_CLI_RC = u'MSSQL_CLI_RC'
 
-from mssqlcli.i18n import translation
-translation().install()
 
 def create_parser():
     args_parser = argparse.ArgumentParser(
@@ -24,130 +22,130 @@ def create_parser():
         dest=u'username',
         default=os.environ.get(MSSQL_CLI_USER, None),
         metavar=u'',
-        help=_(u'Username to connect to the database'))
+        help=u'Username to connect to the database')
 
     args_parser.add_argument(
         u'-P', u'--password',
         dest=u'password',
         default=os.environ.get(MSSQL_CLI_PASSWORD, None),
         metavar=u'',
-        help=_(u'If not supplied, defaults to value in environment variable MSSQL_CLI_PASSWORD.'))
+        help=u'If not supplied, defaults to value in environment variable MSSQL_CLI_PASSWORD.')
 
     args_parser.add_argument(
         u'-d', u'--database',
         dest=u'database',
         default=os.environ.get(MSSQL_CLI_DATABASE, None),
         metavar=u'',
-        help=_(u'database name to connect to.'))
+        help=u'database name to connect to.')
 
     args_parser.add_argument(
         u'-S', u'--server',
         dest=u'server',
         default=os.environ.get(MSSQL_CLI_SERVER, u'localhost'),
         metavar=u'',
-        help=_(u'SQL Server instance name or address.'))
+        help=u'SQL Server instance name or address.')
 
     args_parser.add_argument(
         u'-E', u'--integrated',
         dest=u'integrated_auth',
         action=u'store_true',
         default=False,
-        help=_(u'Use integrated authentication on windows.'))
+        help=u'Use integrated authentication on windows.')
 
     args_parser.add_argument(
         u'-v', u'--version',
         dest=u'version',
         action=u'store_true',
         default=False,
-        help=_(u'Version of mssql-cli.'))
+        help=u'Version of mssql-cli.')
 
     args_parser.add_argument(
         u'--mssqlclirc',
         dest=u'mssqlclirc_file',
         default=os.environ.get(MSSQL_CLI_RC, config_location() + 'config'),
         metavar=u'',
-        help=_(u'Location of mssqlclirc config file.'))
+        help=u'Location of mssqlclirc config file.')
 
     args_parser.add_argument(
         u'--row-limit',
         dest=u'row_limit',
         default=os.environ.get(MSSQL_CLI_ROW_LIMIT, None),
         metavar=u'',
-        help=_(u'Set threshold for row limit prompt. Use 0 to disable prompt.'))
+        help=u'Set threshold for row limit prompt. Use 0 to disable prompt.')
 
     args_parser.add_argument(
         u'--less-chatty',
         dest=u'less_chatty',
         action=u'store_true',
         default=False,
-        help=_(u'Skip intro on startup and goodbye on exit.'))
+        help=u'Skip intro on startup and goodbye on exit.')
 
     args_parser.add_argument(
         u'--auto-vertical-output',
         dest=u'auto_vertical_output',
         action=u'store_true',
         default=False,
-        help=_(u'Automatically switch to vertical output mode if the result is wider than the terminal width.'))
+        help=u'Automatically switch to vertical output mode if the result is wider than the terminal width.')
 
     args_parser.add_argument(
         u'-N', u'--encrypt',
         dest=u'encrypt',
         action=u'store_true',
         default=False,
-        help=_(u'SQL Server uses SSL encryption for all data if the server has a certificate installed.'))
+        help=u'SQL Server uses SSL encryption for all data if the server has a certificate installed.')
 
     args_parser.add_argument(
         u'-C', u'--trust-server-certificate',
         dest=u'trust_server_certificate',
         action=u'store_true',
         default=False,
-        help=_(u'The channel will be encrypted while bypassing walking the certificate chain to validate trust.'))
+        help=u'The channel will be encrypted while bypassing walking the certificate chain to validate trust.')
 
     args_parser.add_argument(
         u'-l', u'--connect-timeout',
         dest=u'connection_timeout',
         default=0,
         metavar=u'',
-        help=_(u'Time in seconds to wait for a connection to the server before terminating request.'))
+        help=u'Time in seconds to wait for a connection to the server before terminating request.')
 
     args_parser.add_argument(
         u'-K', u'--application-intent',
         dest=u'application_intent',
         metavar=u'',
-        help=_(u'Declares the application workload type when connecting to a database in a SQL Server Availability Group.'))
+        help=u'Declares the application workload type when connecting to a database in a SQL Server Availability Group.')
 
     args_parser.add_argument(
         u'-M', u'--multi-subnet-failover',
         dest=u'multi_subnet_failover',
         action=u'store_true',
         default=False,
-        help=_(u'If application is connecting to AlwaysOn AG on different subnets, setting this provides faster detection and connection to currently active server.'))
+        help=u'If application is connecting to AlwaysOn AG on different subnets, setting this provides faster detection and connection to currently active server.')
 
     args_parser.add_argument(
         u'-a', u'--packet-size',
         dest=u'packet_size',
         default=0,
         metavar=u'',
-        help=_(u'Size in bytes of the network packets used to communicate with SQL Server.'))
+        help=u'Size in bytes of the network packets used to communicate with SQL Server.')
 
     args_parser.add_argument(
         u'-A', u'--dac-connection',
         dest=u'dac_connection',
         action=u'store_true',
         default=False,
-        help=_(u'Connect to SQL Server using the dedicated administrator connection.'))
+        help=u'Connect to SQL Server using the dedicated administrator connection.')
 
     args_parser.add_argument(
         u'--enable-sqltoolsservice-logging',
         dest=u'enable_sqltoolsservice_logging',
         action=u'store_true',
         default=False,
-        help=_(u'Enables diagnostic logging for the SqlToolsService.'))
+        help=u'Enables diagnostic logging for the SqlToolsService.')
 
     args_parser.add_argument(
         u'--prompt',
         dest=u'prompt',
         metavar=u'',
-        help=_(u'Prompt format (Default: \\d> '))
+        help=u'Prompt format (Default: \\d> ')
 
     return args_parser
